fix(renderer): make the exact CPU path architecture-independent and enable the ARM64 rows (Task 10.20) (#91)
* fix(renderer): make the exact CPU path architecture-independent (Task 10.20)

The renderer's float64 arithmetic relied on the Go backend contracting each
multiply-add into one fused operation on ARM64. That matched the NEON span
kernel but matched neither amd64 nor the correctness oracle, which rounds with
math.Round and cannot fuse, so `internal/fit/renderer` was excluded from both
ARM64 rows of `ci-native-simd.yml`. Both rows now run it.

The recorded diagnosis was wrong on three counts, and the corrected one is in
`docs/known-limitations.md`:

- It named `compositeScalarPixel`, which does not exist; the alpha store is in
  `compositePixel`. Blocking that store alone changes nothing, because the
  oracle computes `outputAlpha` with the same expression shape and fuses with
  it.
- It said the failure needed ARM64 hardware. `qemu-aarch64-static` plus binfmt
  runs the cross-compiled test binary directly, reports NEON, and reproduces
  `pixel (4,11) channel 3 = 205, baseline = 206` exactly.
- It recorded one failing test where there are two. `TestPolishCircleBatchPool-
  WidthParity` also fails, against expected costs recorded on amd64.

Compiling the unmodified tree with `-gcflags=all=-d=fmahash=<pattern>` makes the
whole ARM64 suite pass, which bounds the cause to contraction alone, and
`bisect -compile=fma` named `composite_span.go:19:24` as the decisive site.

Two traps are worth naming. An explicit conversion rounds only what it wraps, so
`fgR + float64(bgR*bgBlend)` still fuses against `r*alpha` unless the
premultiplied foreground is rounded too. And the scalar span compositor was
deliberately fused to match the NEON kernel, so the kernel had to be unfused
with it: a 1:1 swap of `VMOV`+`VFMLA` for `FMUL`+`FADD`, hand-encoded because
Go's arm64 assembler has no vector mnemonic for them. Each encoding's comment is
what `go tool objdump` prints for it.

Nothing had ever compared that kernel against its scalar reference - both ARM64
span tests force the scalar tier, so they compared the scalar path with itself.
`TestCompositeSpanNEONMatchesScalar` now does, and its k/255 colour sampling is
load-bearing: uniformly random float colours found zero mismatches in 51.2
million evaluations against a reference that was genuinely wrong, where k/255
colours expose it in about half a percent of bytes. The test fails against the
previous fused kernel and passes against this one.

The span-geometry conversions are defensive rather than required: the suite
passes without them. They close the same class of divergence in the coverage
decision, once per row rather than per pixel.

amd64 output is unchanged - it emits no VFMADD before or after - and a 640x480
scene with NEON-length spans now renders to the same bytes and the same cost
under arm64/NEON and amd64/AVX2.

Verified: gofmt, go vet, go build, go test -short ./..., go test -race -short
./internal/fit/..., scripts/check-cross-build.sh (all six targets), golangci-lint
--new-from-merge-base=origin/main (0 issues), and the renderer and fit suites
under qemu-aarch64 at the NEON and forced-scalar tiers, including the
CIRCLEFIT_REQUIRE_SIMD_TIER combinations the workflow uses. ARM64 throughput was
not measured; emulated timings would not count, and `compositeSpanNEONMinPixels`
is unchanged.


* test(renderer): derive parallel-evaluation counts from GOMAXPROCS

Three tests in internal/fit/renderer hardcoded a thread or worker count and
asserted it came back unchanged. Both setters clamp to GOMAXPROCS, so those
assertions were a hidden requirement for a four-processor machine.

Nothing noticed because the package only ever ran on four-core Linux amd64
runners. Enabling it on the ARM64 rows exposed it immediately: the macOS ARM64
runner has three processors, and the first CI run of this branch failed there
with `pool width = 3, want 4` while Linux ARM64 passed.

The counts are now `min(requested, GOMAXPROCS(0))`, matching the idiom the same
file already uses for ParallelEvaluationWorkers, so the tests keep their
coverage on a smaller machine instead of being skipped. The package now passes
at GOMAXPROCS 1, 2, 3 and 4; on main it fails at 1 and at 3.

This is unrelated to the multiply-add contraction fix in the previous commit -
the failures are CPU-count assumptions, not arithmetic - but it is what the
first non-emulated ARM64 run turned up.


* fix(renderer): round each squared radius at its assignment

Addresses the review on #91. The span-geometry conversions added in 027a2f1
rounded only `dy*dy`, which does not block anything: `radiusSquared` is itself a
product, so arm64 fused it straight into the subtraction as FNMSUBD and the
codegen still showed a fused op at every one of those lines. A span boundary
could therefore land differently than on amd64, and the incremental dirty mask
could include or omit a pixel that the full renderer did not.

Each squared radius is now rounded where it is assigned, in both the float64 and
float32 geometry paths and in the full-render scanline walker. No fused
multiply-add is left on any path that decides a pixel.

Four sites are left fused on purpose, because they choose search behaviour
rather than output: the circle-area budget heuristics in polish_dirty_cost.go
and incremental_cost.go, and the seed separation test and correctiveChannel in
batch_audit.go. docs/known-limitations.md records that, and that it means fits
are not bit-reproducible across architectures even though renders now are.

`r2` is renamed to `radiusSquared` in renderCircleScanlineRowsTracked because
adding the comment moved its declaration and varnamelen then reported it.

AGENTS.md said the ARM64 rows skip internal/fit/renderer and that
check-cross-build.sh is the only thing that catches ARM64 test-file problems.
Both were true when written and neither is now, so it is updated: the script is
the only *local* gate, CI covers it at run time, and the macOS ARM64 runner's
three processors are called out so the next hardcoded thread count is caught
before it is pushed.

Verified: gofmt, go vet, go build, go test -short ./..., go test -race -short
./internal/fit/..., check-cross-build.sh (six targets), golangci-lint
--new-from-merge-base=origin/main (0 issues), the renderer suite under
qemu-aarch64, and amd64 at GOMAXPROCS 1, 3 and 4.


---------

+++ internal/fit/renderer/composite_span.go
@@ -1,13 +1,30 @@
 package renderer
 
 // compositeOpaqueSpanScalar composites pixels from an opaque NRGBA span. The
-// arithmetic intentionally matches compositePixel's opaque-destination path;
-// in particular, keeping each multiply-add in one expression lets ARM64 use
-// the same fused operations as the NEON implementation.
+// arithmetic intentionally matches compositePixel's opaque-destination path.
+//
+// The float64 conversions are load-bearing and must not be deleted as
+// redundant. Go allows an implementation to contract a*b+c into one fused
+// multiply-add that rounds once, and the ARM64 backend takes that option while
+// amd64 does not, so the same source produced different bytes on the two
+// targets. Every explicit conversion here is a rounding point the spec forbids
+// contraction from crossing, which is what makes this function architecture-
+// independent. Note that fgR/fgG/fgB have to be rounded too: they are products
+// themselves, so without the conversion the compiler fuses the outer add
+// against r*alpha instead of against bgR*bgBlend.
+//
+// An earlier version of this comment said the opposite - that keeping each
+// multiply-add in one expression was deliberate, so ARM64 would fuse the same
+// way the NEON kernel did. That made the scalar path match the kernel at the
+// cost of matching neither amd64 nor the correctness oracle, and it is why
+// internal/fit/renderer could not run on the ARM64 CI rows. The NEON kernel is
+// now unfused to match this function; TestCompositeSpanNEONMatchesScalar pins
+// that, and composite_span_fast.go documents the separate opt-in path where
+// architecture-dependent rounding is still accepted.
 func compositeOpaqueSpanScalar(pix []byte, offset, pixels int, r, g, b, alpha float64) {
-	fgR := r * alpha
-	fgG := g * alpha
-	fgB := b * alpha
+	fgR := float64(r * alpha)
+	fgG := float64(g * alpha)
+	fgB := float64(b * alpha)
 	bgBlend := 1 - alpha
 
 	end := offset + pixels*4
@@ -15,9 +32,9 @@ func compositeOpaqueSpanScalar(pix []byte, offset, pixels int, r, g, b, alpha fl
 		bgR := float64(pix[i+0]) * inv255
 		bgG := float64(pix[i+1]) * inv255
 		bgB := float64(pix[i+2]) * inv255
-		pix[i+0] = uint8((fgR+bgR*bgBlend)*255 + 0.5)
-		pix[i+1] = uint8((fgG+bgG*bgBlend)*255 + 0.5)
-		pix[i+2] = uint8((fgB+bgB*bgBlend)*255 + 0.5)
+		pix[i+0] = uint8(float64((fgR+float64(bgR*bgBlend))*255) + 0.5)
+		pix[i+1] = uint8(float64((fgG+float64(bgG*bgBlend))*255) + 0.5)
+		pix[i+2] = uint8(float64((fgB+float64(bgB*bgBlend))*255) + 0.5)
 	}
 }
 
@@ -26,9 +43,9 @@ func compositeOpaqueSpanScalar(pix []byte, offset, pixels int, r, g, b, alpha fl
 // the blend setup and loop control exposes two independent pixel streams to
 // the CPU without changing the per-pixel arithmetic.
 func compositeOpaqueSpanPairScalar(pix []byte, firstOffset, secondOffset, pixels int, r, g, b, alpha float64) {
-	fgR := r * alpha
-	fgG := g * alpha
-	fgB := b * alpha
+	fgR := float64(r * alpha)
+	fgG := float64(g * alpha)
+	fgB := float64(b * alpha)
 	bgBlend := 1 - alpha
 
 	firstEnd := firstOffset + pixels*4
@@ -38,16 +55,16 @@ func compositeOpaqueSpanPairScalar(pix []byte, firstOffset, secondOffset, pixels
 		bgR := float64(pix[first+0]) * inv255
 		bgG := float64(pix[first+1]) * inv255
 		bgB := float64(pix[first+2]) * inv255
-		pix[first+0] = uint8((fgR+bgR*bgBlend)*255 + 0.5)
-		pix[first+1] = uint8((fgG+bgG*bgBlend)*255 + 0.5)
-		pix[first+2] = uint8((fgB+bgB*bgBlend)*255 + 0.5)
+		pix[first+0] = uint8(float64((fgR+float64(bgR*bgBlend))*255) + 0.5)
+		pix[first+1] = uint8(float64((fgG+float64(bgG*bgBlend))*255) + 0.5)
+		pix[first+2] = uint8(float64((fgB+float64(bgB*bgBlend))*255) + 0.5)
 
 		bgR = float64(pix[second+0]) * inv255
 		bgG = float64(pix[second+1]) * inv255
 		bgB = float64(pix[second+2]) * inv255
-		pix[second+0] = uint8((fgR+bgR*bgBlend)*255 + 0.5)
-		pix[second+1] = uint8((fgG+bgG*bgBlend)*255 + 0.5)
-		pix[second+2] = uint8((fgB+bgB*bgBlend)*255 + 0.5)
+		pix[second+0] = uint8(float64((fgR+float64(bgR*bgBlend))*255) + 0.5)
+		pix[second+1] = uint8(float64((fgG+float64(bgG*bgBlend))*255) + 0.5)
+		pix[second+2] = uint8(float64((fgB+float64(bgB*bgBlend))*255) + 0.5)
 	}
 }
